complete removal of deprecated classes
