--- java/test/jmri/managers/configurexml/InternalLightManagerXmlTest.java
@@ -1,34 +1,0 @@
-package jmri.managers.configurexml;
-
-import jmri.util.JUnitUtil;
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- *
- * @author Paul Bender Copyright (C) 2017	
- */
-public class InternalLightManagerXmlTest {
-
-    @Test
-    public void testCTor() {
-        InternalLightManagerXml t = new InternalLightManagerXml();
-        Assert.assertNotNull("exists",t);
-    }
-
-    // The minimal setup for log4J
-    @Before
-    public void setUp() {
-        JUnitUtil.setUp();
-    }
-
-    @After
-    public void tearDown() {
-        JUnitUtil.tearDown();
-    }
-
-    // private final static Logger log = LoggerFactory.getLogger(InternalLightManagerXmlTest.class);
-
-}
