+++ java/test/jmri/managers/configurexml/PackageTest.java
@@ -15,9 +15,6 @@ import org.junit.runners.Suite;
 	DefaultSignalMastLogicManagerXmlTest.class,
 	DefaultSignalMastManagerXmlTest.class,
 	DefaultUserMessagePreferencesXmlTest.class,
-	InternalLightManagerXmlTest.class,
-	InternalReporterManagerXmlTest.class,
-	InternalTurnoutManagerXmlTest.class,
 	ManagerDefaultSelectorXmlTest.class,
 	ProxyTurnoutManagerXmlTest.class
 })
